moved helium nlte to a seperate file
